ref(evals): Move legacy code under legacy tree
Relocate the remaining scorer and evaluate implementations under src/legacy so the internal layout matches the new root-versus-legacy public API split.

Keep judge wrappers pointing at the relocated legacy code and move the corresponding tests with the implementation so the compatibility surface stays intact while the new harness-first API continues to live at the root.

Refs #39
