+++ packages/vitest-evals/src/legacy/evaluate/index.ts
@@ -1,7 +1,7 @@
 import { generateObject } from "ai";
 import { z } from "zod";
 import { assert, test } from "vitest";
-import { wrapText } from "../wrapText";
+import { wrapText } from "../../wrapText";
 
 type LanguageModel = Parameters<typeof generateObject>[0]["model"];
