+++ packages/vitest-evals/src/legacy/scorers/structuredOutputScorer.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect, vi } from "vitest";
-import "../legacy"; // Import to register the toEval matcher
+import "../../legacy"; // Import to register the toEval matcher
 import { StructuredOutputScorer } from "./structuredOutputScorer";
 
 describe("StructuredOutputScorer", () => {
