+++ packages/vitest-evals/src/legacy/scorers/structuredOutputScorer.ts
@@ -1,4 +1,4 @@
-import type { BaseScorerOptions, ScoreFn } from "../legacy/shared";
+import type { BaseScorerOptions, ScoreFn } from "../shared";
 import {
   type BaseMatcherConfig,
   type MatchStrategy,
